V1.32 11.08.2017 S10history integriert
- S10history von [RalfJL](https://github.com/RalfJL/S10history)
integriert
- [Anleitunf für
S10History](https://github.com/nischram/E3dcGui/tree/master/S10history)
angepasst
- Abfrage der HistoryValues für Tag und Vortag
- Anzeige der HistoryValues für Tag und Vortag in den Aktuellen Werten
- Issue #12
- Anpassungen im RSCP für S10history
- ASE.cpp für 64Bit vorbereitet Issue #4
- Issue #15 WatchDog verbessert
- IP Abfrage für das Setup Menü angepasst
- Funktionen für Setup Menü ausgelagert !!! Achtung !!! In der
screenSaveHM.c muss folgende Zeile im oberen Bereich eingefügt werden:
`  #include <sys/sysinfo.h>  `
- Zusammenfassen von Funktionen in der GuiMain.c

+++ Rscp/AES.h
@@ -33,6 +33,8 @@
 #ifndef _AES_H
 #define _AES_H
 
+#include <stdint.h>
+
 /* USAGE:
 	1. Create a AES class (or more as necessary)
 	2. Call class method SetParameters
@@ -91,7 +93,7 @@ public:
 	// have enough space in datain and dataout to accomodate this. Pad your data before
 	// calling, preferably using the padding methods listed below.
 	// Decryption must use the same mode as the encryption.
-	void Encrypt(const unsigned char * datain, unsigned char * dataout, unsigned long numBlocks, BlockMode mode = CBC);
+	void Encrypt(const unsigned char * datain, unsigned char * dataout, uint32_t numBlocks, BlockMode mode = CBC);
 
 	// call this before any decryption with the key to use
 	void StartDecryption(const unsigned char * key);
@@ -102,12 +104,12 @@ public:
 	// calling, preferably using the padding methods listed below. You must know the desired
 	// length of the output data, since all the blocks are returned decrypted.
 	// Encryption must use the same mode as the decryption.
-	void Decrypt(const unsigned char * datain, unsigned char * dataout, unsigned long numBlocks, BlockMode mode = CBC);
+	void Decrypt(const unsigned char * datain, unsigned char * dataout, uint32_t numBlocks, BlockMode mode = CBC);
 
 private:
 
-	int Nb,Nk;    // block and key length / 32, should be 4,6,or 8
-	int Nr;       // number of rounds
+	int32_t Nb,Nk;    // block and key length / 32, should be 4,6,or 8
+	int32_t Nr;       // number of rounds
 
 	unsigned char W[4*8*15];   // the expanded key
 	unsigned char iv[32];  	   // initial value which is incremented
